Split customers schema file in directory
