+++ customers/schema/__init__.py
@@ -1,0 +1,25 @@
+from .mutations import Mutation
+from .queries import Query
+from .types import (
+    BerthProfileNode,
+    BoatCertificateNode,
+    BoatCertificateTypeEnum,
+    BoatNode,
+    InvoicingTypeEnum,
+    OrganizationNode,
+    OrganizationTypeEnum,
+    ProfileNode,
+)
+
+__all__ = [
+    "BerthProfileNode",
+    "BoatCertificateNode",
+    "BoatCertificateTypeEnum",
+    "BoatNode",
+    "InvoicingTypeEnum",
+    "Mutation",
+    "OrganizationNode",
+    "OrganizationTypeEnum",
+    "ProfileNode",
+    "Query",
+]
